Add movingpandas import
