+++ lonboard/_layer.py
@@ -19,7 +19,6 @@ from lonboard.traits import (
     ColorAccessor,
     FloatAccessor,
     PyarrowTableTrait,
-    TimestampAccessor,
 )
 
 if TYPE_CHECKING:
@@ -967,148 +966,6 @@ class HeatmapLayer(BaseArrowLayer):
     """
 
     @traitlets.validate("get_weight")
-    def _validate_accessor_length(self, proposal):
-        if isinstance(proposal["value"], (pa.ChunkedArray, pa.Array)):
-            if len(proposal["value"]) != len(self.table):
-                raise traitlets.TraitError("accessor must have same length as table")
-
-        return proposal["value"]
-
-
-class TripsLayer(BaseArrowLayer):
-    """ """
-
-    _layer_type = traitlets.Unicode("trip").tag(sync=True)
-
-    table = PyarrowTableTrait(
-        allowed_geometry_types={
-            EXTENSION_NAME.LINESTRING,
-        }
-    )
-
-    width_units = traitlets.Unicode(allow_none=True).tag(sync=True)
-    """
-    The units of the line width, one of `'meters'`, `'common'`, and `'pixels'`. See
-    [unit
-    system](https://deck.gl/docs/developer-guide/coordinate-systems#supported-units).
-
-    - Type: `str`, optional
-    - Default: `'meters'`
-    """
-
-    width_scale = traitlets.Float(allow_none=True, min=0).tag(sync=True)
-    """
-    The path width multiplier that multiplied to all paths.
-
-    - Type: `float`, optional
-    - Default: `1`
-    """
-
-    width_min_pixels = traitlets.Float(allow_none=True, min=0).tag(sync=True)
-    """
-    The minimum path width in pixels. This prop can be used to prevent the path from
-    getting too thin when zoomed out.
-
-    - Type: `float`, optional
-    - Default: `0`
-    """
-
-    width_max_pixels = traitlets.Float(allow_none=True, min=0).tag(sync=True)
-    """
-    The maximum path width in pixels. This prop can be used to prevent the path from
-    getting too thick when zoomed in.
-
-    - Type: `float`, optional
-    - Default: `None`
-    """
-
-    joint_rounded = traitlets.Bool(allow_none=True).tag(sync=True)
-    """
-    Type of joint. If `True`, draw round joints. Otherwise draw miter joints.
-
-    - Type: `bool`, optional
-    - Default: `False`
-    """
-
-    cap_rounded = traitlets.Bool(allow_none=True).tag(sync=True)
-    """
-    Type of caps. If `True`, draw round caps. Otherwise draw square caps.
-
-    - Type: `bool`, optional
-    - Default: `False`
-    """
-
-    miter_limit = traitlets.Int(allow_none=True).tag(sync=True)
-    """
-    The maximum extent of a joint in ratio to the stroke width.
-    Only works if `jointRounded` is `False`.
-
-    - Type: `float`, optional
-    - Default: `4`
-    """
-
-    billboard = traitlets.Bool(allow_none=True).tag(sync=True)
-    """
-    If `True`, extrude the path in screen space (width always faces the camera).
-    If `False`, the width always faces up.
-
-    - Type: `bool`, optional
-    - Default: `False`
-    """
-
-    fade_trail = traitlets.Bool(allow_none=True).tag(sync=True)
-    """Whether or not the path fades out.
-
-    - Type: `bool`, optional
-    - Default: `True`
-    """
-
-    trail_length = traitlets.Float(allow_none=True).tag(sync=True)
-    """Trail length.
-
-    - Type: `float`, optional
-    - Default: `120`
-    """
-
-    current_time = traitlets.Float(0).tag(sync=True)
-    """The current time of the frame.
-
-    - Type: `float`, optional
-    - Default: `0`
-    """
-
-    get_color = ColorAccessor()
-    """
-    The color of each path in the format of `[r, g, b, [a]]`. Each channel is a number
-    between 0-255 and `a` is 255 if not supplied.
-
-    - Type: [ColorAccessor][lonboard.traits.ColorAccessor], optional
-        - If a single `list` or `tuple` is provided, it is used as the color for all
-          paths.
-        - If a numpy or pyarrow array is provided, each value in the array will be used
-          as the color for the path at the same row index.
-    - Default: `[0, 0, 0, 255]`.
-    """
-
-    get_width = FloatAccessor()
-    """
-    The width of each path, in units specified by `width_units` (default `'meters'`).
-
-    - Type: [FloatAccessor][lonboard.traits.FloatAccessor], optional
-        - If a number is provided, it is used as the width for all paths.
-        - If an array is provided, each value in the array will be used as the width for
-          the path at the same row index.
-    - Default: `1`.
-    """
-
-    get_timestamps = TimestampAccessor()
-    """
-    The timestamp of each coordinate.
-
-    - Type: [TimestampAccessor][lonboard.traits.TimestampAccessor]
-    """
-
-    @traitlets.validate("get_color", "get_width", "get_timestamps")
     def _validate_accessor_length(self, proposal):
         if isinstance(proposal["value"], (pa.ChunkedArray, pa.Array)):
             if len(proposal["value"]) != len(self.table):
