+++ lonboard/experimental/__init__.py
@@ -4,5 +4,5 @@ These layers have not been as well tested as other layers. You may encounter cra
 unexpected behavior when using them.
 """
 
-from ._layer import ArcLayer, TextLayer
+from ._layer import ArcLayer, TextLayer, TripsLayer
 from .layer_extension import BrushingExtension, CollisionFilterExtension
