+++ lonboard/experimental/traits.py
@@ -9,6 +9,7 @@ from traitlets.traitlets import TraitType
 
 from lonboard._serialization import (
     COLOR_SERIALIZATION,
+    FLOAT_SERIALIZATION,
 )
 from lonboard.traits import FixedErrorTraitType
 
@@ -98,6 +99,43 @@ class PointAccessor(FixedErrorTraitType):
                 return
 
             return tuple(map(int, (np.array(c) * 255).astype(np.uint8)))
+
+        self.error(obj, value)
+        assert False
+
+
+class TimestampAccessor(FixedErrorTraitType):
+    """A representation of a deck.gl coordinate-timestamp accessor.
+
+    - A pyarrow [`ListArray`][pyarrow.ListArray] containing either a `FloatArray` or
+      `DoubleArray`. Each value in the array will be used as the value for the object at
+      the same row index.
+    """
+
+    default_value = None
+    info_text = (
+        "a float value or numpy ndarray or pyarrow array representing an array"
+        " of floats"
+    )
+
+    def __init__(
+        self: TraitType,
+        *args,
+        **kwargs: Any,
+    ) -> None:
+        super().__init__(*args, **kwargs)
+        self.tag(sync=True, **FLOAT_SERIALIZATION)
+
+    def validate(self, obj, value) -> Union[float, pa.ChunkedArray, pa.DoubleArray]:
+        if isinstance(value, (pa.ChunkedArray, pa.Array)):
+            if not pa.types.is_list(value.type):
+                self.error(
+                    obj,
+                    value,
+                    info="pyarrow array must be a list type.",
+                )
+
+            return value
 
         self.error(obj, value)
         assert False
